React to expected changes in the cluster
This takes into account the annotations that the MCO adds to the nodes
in order to react to changes that are "approved" or expected in the
cluster.

Specifically, if the node is expecting changes (currentConfig !=
desiredConfig and the MCD is working) we create a small workload that
creates a file that's read by the AIDE pod. This file tells the AIDE pod
to hold off doing integrity checks.

If the node is done doing changes or it hasn't changed (currentConfig ==
desiredConfig and the MCD is done) we remove the workload that adds the
hold-off file, and create another one that removes that file. This tells
the AIDE pod to continue doing integrity checks as usual.

Co-Authored-By: EvB <evb@redhat.com>
Co-Authored-By: Antonio Murdaca <runcom@linux.com>